refactored to two files and initialized NtiveStackNavigator on entrypoint file
