+++ frontend/documentation/development_diary.txt
@@ -1,3 +1,9 @@
 Entry 1 - Mar 8
 - Using this page to help implement navigation through the app https://reactnative.dev/docs/navigation#react-navigation
 - TODO: figure out questions about cocoapods and how that works
+
+Entry 2
+- Worked to refactor files into two apps to facilitate dedicating App.js to the NavigationContainer
+- imported components from folder
+- initilized the objects amnd tags that deal with navigation
+- Needed to run this https://reactnavigation.org/docs/troubleshooting/#im-getting-an-error-unable-to-resolve-module-after-updating-to-the-latest-version to run on the simulator
